Employee - Optimaize duplicate code
