+++ components/modules/employees/index.ts
@@ -1,1 +1,2 @@
 export * from './Detail';
+export * from './EmployeeForm';
